Add Public Dandisets View
Add the first draft of the public dandisets view. I anticipate the
following changes at some point:

* UX styling for the list dandisets component. It is currently not
filling the page and still needs some more work done to get all the
stats it needs.
* Comprehensive search solution. I included a placeholder `SearchField`
that looks ok, but does nothing. When search is implemented properly it
should be replaced with a general component.
* Better search option chips. We don't have the information available to
do all of Jared's recommended chips, and the UX of what I did is
probably unsatisfactory in some way. It is fully functional and
easily fixable though, so I think it's OK for an MVP.