RF: nwb.json => dandiset_metanwb.json
Initially I had missed the fact that it is not the schema we impose on top
of nwb, but rather a schema for dandiset.yaml fields which are derived from
metadata we find in nwb files.  As it is not directly the information from
nwb, I decided to call it metanwb.

This is a follow up to previous renamings in 9acd9d07b06d45b617ccc37ac729e5950ce1d819