Add more end-to-end tests
Add the following tests:
* registerDandiset - Tests creating a new dandiset
* stats - Tests that stats increment correctly
* search - Tests the basic search functionality

Also add a new flavor of helper, pages. Pages are stored in
`test/src/pages/`. The contain methods that interact with a certain view
in the app. For example, `homePage` contains the methods `getStat` and
`search`, as those are the only two actions available on the home page.
The methods in `searchPage` will need to be fleshed out as we add more
tests for search filtering/paging.