Add a locking service to girder
Add two endpoints, `dandi/:identifier/lock` and
`dandi/:identifier/unlock`. These will lock or unlock dandisets for the
current user. When unlocked, a dandiset can be modified by anyone with
permission. When locked, a dandiset can be modified only by the person
who owns the lock.

Locks are enforced using event listeners on folder save and assetstore
upload events. Modifying a dandiset's metadata or uploading new files to
it will fail if another user has locked the dandiset.

The current lock status is stored in a property on the root folder
model. We may eventually want to add an endpoint to return the current
lock status so it can be displayed in the web client.