Make sync synchronous instead of celery task
The sync action was originally a celery task so that the API did not
have to block on any database operations. This is inconvenient when
syncing immediately before a publish. The sync request would return, but
the celery task was still queued, so immediately performing a publish
would fail because the database had not yet been synced.

The sync endpoint now blocks until the sync is complete.