Remove isPublishedVersion util function
This is because since the function was added, the use of published
versions has evolved, and now the name is misleading (as it implies the
version specified is actually a valid published version, which is not
what the function evaluated). Furthermore, the function was only useful
some of the time, and could cause more confusion.

This also adds the dandisetHasVersion function, which accomplishes
something the removed function did not.