Add Draft and DraftAsset models
DraftAssets are currently not plugged in to anything, they just exist to
parallel Asset.

Creating a new dandiset from girder will also create a new Draft with
the same content, for populating Drafts with.

There are two migrations included, one to modify the schema and one to
generate a Draft for the latest publish Version for any dandisets.