Collapse nav toolbar into dropdown when <= md
Adding the documentation link to the toolbar was causing the login
avatar to be shoved off the edge off of the screen when automated tests
ran.

Rather than fix the screen size, the app should behave intelligently for
smaller screen sizes (i.e. mobile). The toolbar now collapses into a
dropdown menu when the screensize is medium is smaller.