Save AssetBlob after validation succeeds
In addition to setting the validation state to `SUCCEEDED`, an AssetBlob
is saved. This will eventually involve copying between storages, so this
must be done in the validation celery task.

Now creating an asset does not involve checking validations, as an
AssetBlob will always exist with the sha256 checksum if the validation
succeeded.