Clean up asset create/update views
The Asset `create`/`update` views now use the same serializer and an almost
identical flow. The only difference is that `create` will fail if the
asset already exists, and `update` will fail if the asset does not yet
exist.

Assets are now identified in the API by their UUID, not their path.