Factor the dialog component out of the items v-for
This was causing a bug where a single v-model value was controlling
several <v-dialog>s, and all of them would open when any delete button
was clicked. This had two modes of badness: first, the dialog for the
last item in the list would show up on top, confirming the wrong asset
in general; second, obviously, we should not be rendering N dialogs for
N items.