Do not override DJANGO_DEFAULT_FROM_EMAIL
The from address was hardcoded when sending messages, which broke
message sending on the staging instance. It should instead use the
default Django value, which is set appropriately in Teraform.

Also add a print statement to the user registration signal handler so
there is some debug output when a user signs up.