Add workaround to permit localhost URLs
The default dandischema pydantic model does not allow localhost in URLs,
which is a requirement for test and dev environments. To get around
this, the dandischema module will subtly tweak the model to be more
permissive with URLs if the environment variable
DANDI_ALLOW_LOCALHOST_URLS is set.

See https://github.com/dandi/dandischema/pull/33