Clear the ownership search field after selection
The normal behavior for Vuetify autocompletes is (reasonably) to put the
selected text into the search field. We want to add the selected user to
a list, then clear the search field.

This also fixes a 400 error where the dialog box would search for the
pretty-printed username and find nothing.