Replace validate_version_metadata.delay() with version.status=PENDING
With the new scheduled job that checks for Versions that need validation
(i.e. are in status `PENDING`), we no longer need to submit a new task
every time we might want to validate version metadata. Instead, we
simply set `version.status = Version.Status.PENDING` and wait for the
job to pick it up.