Don't validate Assets with no sha256
Previously, we validated Asset metadata immediately after
creating/updating an Asset, even if the sha256 checksum had not been
calculated yet for the blob. This resulted in a validation error related
to a missing digest, but that was shortly addressed when the checksum
finished calculation and revalidated the Asset.

However, this system had a race condition where the Asset creation
validation and the checksum completion validation would run
simultaneously. The checksum validation would save first with the valid
results, then the creation validation would overwrite those results with
erroneous validation errors.

Fix the race condition by not validating after creation/update when the
blob sha256 checksum is not yet calculated.