Don't leak spurious 400 error to client
When the input to user search doesn't validate, it leaks a 400 error
through to the browser (specifically, in the user search form that
appears in the "manage ownership" dialog) but doesn't break user
functionality in any way. In fact, the user typing bad input into this
field *shouldn't* be considered an error, and should instead just return
no results.