Fix zarr publish error bug
`params` must be used to pass query parameters for a GET request with
axios. Since we weren't using that, no dandiset id was being sent to the
zarr endpoint, causing the zarr endpoint to return the count of zarrs
in the entire system, which in turn caused dandisets with no zarrs to be
flagged as having zarrs.