Don't use filesystem APIs to manipulate URLs
Following #1743, this removes the remaining uses of the filesystem APIs
`os.path` and `pathlib` (which are platform-dependent and contain methods for
local filesystem manipulation) to process URLs.

The `zarr_checksum` package still has this confusion in its `ZarrArchiveFile`
class, but attempts to refactor some of that code failed due to other issues
with it. So, the `ZarrArchiveFile.path` field is just cast to a `str` when
interfacing with local application code.