Remove Dandiset add_owner() and remove_owner() methods
This came up during development of the audit functionality. It turns out
that add_owner() was only called during Dandiset creation to add the
creating user as the only owner, and remove_owner() is never called.
Instead, in the method that changes the owner set, assign_perm() is used
in all cases to add and remove owners as needed. Since no other point of
use exists for adding/removing users, it makes sense to remove these
methods.