Explicitly install Python in CD workflows
GitHub Actions' switch to Ubuntu 24.04 breaks the deploy workflows.
We currently use `pip` to install the `build` package prior to packaging
the `dandiapi` sdist. But, we do not explicitly install the Python
runtime before doing this. It looks `pip` was available on the default
Ubuntu 22.04 runners, but not 24.04, so this was basically working by
accident before.