Pass the string value directly to `setLevel()`
The [`logging` module
docs](https://docs.python.org/3/library/logging.html#logging.Logger.setLevel)
explain that `setLevel()` can handle the string representations of the
logging levels. This allows the setting value to be directly tied to
what is passed in, and `setLevel()` also does a "validation" on the
value and will refuse to process levels that don't exist.