Remove `text` prop from `v-list-item`
In Vuetify 2, the `text` prop of the `v-list-item` component is a
boolean that applies a certain styling to the component. In Vuetify 3,
the `text` prop represents the textual content of the `v-list-item`
component. That resulted in it being interpreted as an empty string,
resulting in the issue where those items were not being displayed.